requirements.txt
gymnasium==0.29.1
ale-py==0.9.0
h5py==3.11.0
huggingface-hub==0.17.2
hydra-core==1.3
numpy==1.26.0
opencv-python==4.10.0.84
pillow==10.3.0
pygame==2.5.2
torch==2.4.1
torcheval==0.0.7
tqdm==4.66.4
wandb>=0.17.0
av
ffmpeg-python
nibabel
torchvision==0.19.1
google-cloud-storage
google-cloud-secret-manager
